requirements.txt
pandas
pyarrow
llama-index
gpt-index
openai
pypdf
python-dotenv
typer
python-pptx
requests
html5lib
beautifulsoup4
requests-html